+++ packages/hono-takibi/src/generators/handler/generate-handler-name.ts
@@ -1,4 +1,4 @@
-import { generateRouteName } from '../openapi/paths/generate-route-name'
+import { generateRouteName } from '../hono/openapi/route/generate-route-name'
 
 export function generateHandlerName(method: string, path: string) {
   return `${generateRouteName(method, path)}Handler`
