+++ packages/hono-takibi/src/generators/hono/app/helper/get-route-maps.ts
@@ -1,6 +1,6 @@
 import type { OpenAPISpec } from '../../../../types'
 import { generateHandlerName } from '../../../handler/generate-handler-name'
-import { generateRouteName } from '../../../openapi/paths/generate-route-name'
+import { generateRouteName } from '../../openapi/route/generate-route-name'
 
 export function getRouteMaps(openAPISpec: OpenAPISpec) {
   const paths = openAPISpec.paths
