+++ packages/hono-takibi/src/generators/hono/openapi/route/generate-route-name.ts
@@ -1,4 +1,4 @@
-import { capitalize } from '../../../core/text/capitalize'
+import { capitalize } from '../../../../core/text/capitalize'
 
 /**
  * Generates a route name from HTTP method and path
